requirements.txt
matplotlib>=3.7
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
pytest>=7.0
